restore lost files, use new ab test shape
